refactor(matieres): MatiereDetails sous 300 lignes — éclatement complet (H9)
1174 → 225 lignes. Extraction:
- composables/useMatiereDetails.js (données/logique, route/router via getCurrentInstance lazy pour parité avec le test G10 existant)
- components/matieres/: MatiereNotifications, MatiereDetailsHeader, MatiereLessonsTab (v-model viewMode), CreateLessonModal
- onglets Séances/Évaluations/Classes déjà extraits (#28) consommés tels quels
- overrides Tailwind→thème conservés dans la vue via :deep() (ciblent déjà les descendants)
- tests: composable + 2 sous-composants (13 tests verts, dont les 2 tests G10 existants)

Dette pré-existante conservée: computed canManageVisio non référencé, nombreux console.log de debug.
